+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## rmapi master
+
+*  Fix directory creation (fixes #6)
+
+*  Add stat command to show entry's metadata
+
 ## rmapi 0.0.1 (February 24, 2018)
 
 *   Initial release with support for most of the API and autocompletion.